requirements.txt
pandas
requests
python-dotenv
openpyxl
pyarrow
dash
dash-bootstrap-components
dash-bootstrap-templates
plotly
